prettier: The tragic reformatting associated with 168df742.
Some changes are easier to mentally consume before they've been prettier'd.

While I'll certainly allow Prettier to do what it wants to do, I've confined
it's changes to this commit for easier reviewability.

Note to the reviewer: this commit is exactly the result of running the
`npm run lint-fix` command on a clean Git working tree.

Ref: 168df742a5745bfbd08ccd067bce938aa5a03448