Change types of the Google Cloud Funcs handler to allow options-as-a-function.
In the same manner as AWS Lambda, this now utilizes the
`ServerOptionsFunction` type pattern.  Previously, this typing still appears
to be incorrect, but that incorrectness was being missed because the
sole invocation of this at the end of this integrations `createHandler`
method was utilizing `this.createGraphQLServerOptions.bind(this)`, whose
types were unable to be inferred.  (I'm making a slight assumption on this
last bit because this mis-typing only surfaced after I switched to a more
explicit notation, but I believe this to be true).